requirements.txt
nicegui
pywebview
pymupdf
ollama
ruamel.yaml
pydantic
watchdog
aiosqlite
